requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
ddgs>=9.0.0
wikipedia>=1.4.0
arxiv>=2.1.3
sympy>=1.12
python-dotenv>=1.0.1
deep-translator>=1.11.4
feedparser>=6.0.11
yfinance>=0.2.50
rich>=13.7.1
